rename of style/getlastbill and editing the end point in the client side
